requirements.txt
playwright
behave
allure-behave
pytest-playwright
pytest-check
